Moved all the code of the MMPEVEM to HelpersMMPEVEM.cpp.
