+++ src/utils/emissions/CMakeLists.txt
@@ -1,6 +1,6 @@
-add_subdirectory(MMPEVEM)
-
 set(utils_emissions_STAT_SRCS
+   CharacteristicMap.cpp
+   CharacteristicMap.h
    EnergyParams.cpp
    EnergyParams.h
    HelpersEnergy.cpp
@@ -25,5 +25,4 @@ set(utils_emissions_STAT_SRCS
 )
 
 add_library(utils_emissions STATIC ${utils_emissions_STAT_SRCS})
-target_link_libraries(utils_emissions mmpevem)
 set_property(TARGET utils_emissions PROPERTY PROJECT_LABEL "z_utils_emissions")
