+++ src/utils/emissions/HelpersMMPEVEM.cpp
@@ -6,147 +6,166 @@
  * \copyright Eclipse Public License v2.0
  *            (https://www.eclipse.org/legal/epl-2.0/)
  * \date 2021-10
- * \brief This file contains a helper class for MMPEVEMs.
+ * \brief This file contains MMP's emission model for electric vehicles.
  */
 
+
+#include <utils/emissions/CharacteristicMap.h>
+
+#ifdef _MSC_VER
+#define _USE_MATH_DEFINES
+#endif
+#include <cmath>
 
 #include <utils/emissions/HelpersMMPEVEM.h>
 #include <utils/common/SUMOTime.h>
 #include <utils/common/StringUtils.h>
-
-#include <fstream>
-#include <exception>
-#include <cmath>
-#include <cctype>
+#include <utils/geom/GeomHelper.h>
 
 
 
 
-/**
- * \brief Find a key inside a parameter map and convert its value into a double.
- *
- * \param[in] ref_params A parameter map where both keys and values are strings
- * \param[in] ref_key A key
- *
- * \returns The converted double value
- * \throws std::runtime_error
- */
-double extractDoubleFromParams(
-    const std::map<std::string, std::string>& ref_params,
-    const std::string& ref_key)
-{
-  auto it = ref_params.find(ref_key);
-  if(it == ref_params.end())
-    throw std::runtime_error("Couldn't find the key '" + ref_key + "'.");
-  try
-  {
-    return std::stod(it->second);
-  }
-  catch(const std::exception& ref_e)
-  {
-    throw std::runtime_error("Couldn't convert the value '" + it->second
-        + "' of key '" + ref_key + "' into a double.");
-  }
-}
+// Constants
+const double EPS = 1e-6;
+const double G = 9.81;         // Gravitational acceleration [m/s^2]
+const double RHO_AIR = 1.204;  // Air density [kg/m^3]
+
 
 
 
 /**
- * \brief Find a key inside a parameter map and return its string value.
+ * \brief Compute the power consumption of an EV powertrain in a certain state.
  *
- * \param[in] ref_params A parameter map where both keys and values are strings
- * \param[in] ref_key A key
+ * The model assumes that the driver recuperates as much energy as possible,
+ * meaning that he perfectly employs the mechanical brakes such that only what
+ * lies beyond the motor's capabilities is dissipated. If the new state is
+ * invalid, that is the demanded acceleration exceeds what the electric motor
+ * can manage or the power loss map is not defined for a given point, the torque
+ * and/or power are capped to the motor's limits or the power loss is set to
+ * zero.
  *
- * \returns The value as a string
- * \throws std::runtime_error
+ * \param[in] m Vehicle mass [kg]
+ * \param[in] r_wheel Wheel radius [m]
+ * \param[in] Theta Moment of inertia [kg*m^2]
+ * \param[in] c_rr Rolling resistance coefficient [1]
+ * \param[in] c_d Drag coefficient [1]
+ * \param[in] A_front Cross-sectional area of the front of the car [m^2]
+ * \param[in] i_gear Gear ratio (i_gear = n_motor/n_wheel) [1]
+ * \param[in] eta_gear Gear efficiency [1]
+ * \param[in] M_max Maximum torque of the EM [Nm]
+ * \param[in] P_max Maximum power of the EM [W]
+ * \param[in] M_recup_max Maximum recuperation torque [Nm]
+ * \param[in] P_recup_max Maximum recuperation power [W]
+ * \param[in] R_battery Internal battery resistance [Ohm]
+ * \param[in] U_battery_0 Nominal battery voltage [V]
+ * \param[in] P_const Constant power consumption of ancillary devices [W]
+ * \param[in] ref_powerLossMap Power loss map of the EM + inverter
+ * \param[in] dt Simulation timestep [s]
+ * \param[in] v Vehicle speed at the end of a timestep [m/s]
+ * \param[in] a Constant acceleration during a timestep [m/s^2]
+ * \param[in] alpha Constant incline during a timestep [deg]
+ * \param[out] ref_powerConsumption Power consumption during the last timestep
+ *             [W]
+ * \returns true if the new state is valid, else false
  */
-std::string extractStringFromParams(
-    const std::map<std::string, std::string>& ref_params,
-    const std::string& ref_key)
+bool calcPowerConsumption(double m, double r_wheel, double Theta, double c_rr,
+    double c_d, double A_front, double i_gear, double eta_gear, double M_max,
+    double P_max, double M_recup_max, double P_recup_max, double R_battery,
+    double U_battery_0, double P_const,
+    const CharacteristicMap& ref_powerLossMap, double dt, double v, double a,
+    double alpha, double& ref_powerConsumption)
 {
-  auto it = ref_params.find(ref_key);
-  if(it == ref_params.end())
-    throw std::runtime_error("Couldn't find the key '" + ref_key + "'.");
-  return it->second;
-}
+  bool b_stateValid = true;
 
+  // Mass factor [1]
+  const double e_i = 1.0 + Theta/(m*r_wheel*r_wheel);
+  // Average speed during the previous timestep
+  const double v_mean = v - 0.5*a*dt;
 
+  // Force required for the desired acceleration [N]
+  double F_a = m*a*e_i;
+  // Grade resistance [N]
+  double F_gr = m * G * std::sin(DEG2RAD(alpha));
+  // Rolling resistance [N]
+  double F_rr = m * G * std::cos(DEG2RAD(alpha)) * c_rr;
+  if(std::abs(v_mean) <= EPS)
+    F_rr = 0;
+  // Drag [N]
+  double F_d = 0.5 * c_d * A_front * RHO_AIR * v_mean * v_mean;
+  // Tractive force [N]
+  const double F_tractive = F_a + F_gr + F_rr + F_d;
 
-/**
- * \brief Parse an MMPEVEM model file and create a model base on its parameters.
- *
- * \param[in] ref_modelFile An MMPEVEM model file
- *
- * \returns An MMPEVEM object
- */
-MMPEVEM loadModelFromFile(const std::string& ref_modelFile)
-{
-  std::map<std::string, std::string> params;
+  // Speed of the motor [rpm]
+  const double n_motor = v_mean/(2*M_PI*r_wheel)*60*i_gear;
+  // Angular velocity of the motor [1/s]
+  double omega_motor = 2*M_PI*n_motor/60;
+  if(omega_motor == 0)
+    omega_motor = EPS;
 
-  // Parse the file
-  std::ifstream ifs(ref_modelFile);
-  for(std::string line; std::getline(ifs, line);)
+  // Torque at the motor [Nm]. Please note that this model, like most real EVs,
+  // utilizes the EM to hold the vehicle on an incline rather than the brakes.
+  double M_motor = F_tractive*r_wheel/i_gear;
+  if(F_tractive < 0)
+    M_motor *= eta_gear;
+  else
+    M_motor /= eta_gear;
+  // Engine power [W]
+  double P_motor = M_motor*omega_motor;
+
+  // Cap torque or power if necessary
+  // Accelerating
+  if(M_motor >= 0)
   {
-    // Remove whitespaces
-    line.erase(std::remove_if(line.begin(), line.end(),
-        [](unsigned char x)
-        {
-          return std::isspace(x);
-        }),
-        line.end());
-    // Skip comments
-    if(line[0] == '#')
-      continue;
-
-    // Find the '=' character which delimits keys and values
-    std::size_t delim = line.find("=");
-    if(delim == std::string::npos)
-      continue;
-    std::string k = line.substr(0, delim);
-    std::string v = line.substr(delim + 1);
-    params[k] = v;
+    if(M_motor > M_max)
+    {
+      M_motor = M_max;
+      P_motor = M_motor*omega_motor;
+      b_stateValid = false;
+    }
+    if(P_motor > P_max)
+    {
+      P_motor = P_max;
+      M_motor = P_motor/omega_motor;
+      b_stateValid = false;
+    }
+  }
+  // Recuperating
+  else
+  {
+    // Even when capping, the state is still valid here because the extra energy
+    // is assumed to go to the brakes
+    if(M_motor < -M_recup_max)
+    {
+      M_motor = -M_recup_max;
+      P_motor = M_motor*omega_motor;
+    }
+    if(P_motor < -P_recup_max)
+    {
+      P_motor = -P_recup_max;
+      M_motor = P_motor/omega_motor;
+    }
   }
 
-  // Extract all required parameters
-  // Vehicle mass
-  double m = extractDoubleFromParams(params, "m");
-  // Wheel radius
-  double r_wheel = extractDoubleFromParams(params, "r_wheel");
-  // Internal moment of inertia
-  double Theta = extractDoubleFromParams(params, "Theta");
-  // Rolling resistance coefficient
-  double c_rr = extractDoubleFromParams(params, "c_rr");
-  // Air drag coefficient
-  double c_d = extractDoubleFromParams(params, "c_d");
-  // Cross-sectional area of the front of the car
-  double A_front = extractDoubleFromParams(params, "A_front");
-  // Gear ratio
-  double i_gear = extractDoubleFromParams(params, "i_gear");
-  // Gearbox efficiency
-  double eta_gear = extractDoubleFromParams(params, "eta_gear");
-  // Maximum torque
-  double M_max = extractDoubleFromParams(params, "M_max");
-  // Maximum power
-  double P_max = extractDoubleFromParams(params, "P_max");
-  // Maximum recuperation torque
-  double M_recup_max = extractDoubleFromParams(params, "M_recup_max");
-  // Maximum recuperation power
-  double P_recup_max = extractDoubleFromParams(params, "P_recup_max");
-  // Internal battery resistance
-  double R_battery = extractDoubleFromParams(params, "R_battery");
-  // Nominal battery voltage
-  double U_battery_0 = extractDoubleFromParams(params, "U_battery_0");
-  // Constant power consumption
-  double P_const = extractDoubleFromParams(params, "P_const");
-  // Power loss map
-  std::string powerLossMapString = extractStringFromParams(params,
-      "powerLossMapString");
+  // Power loss in the electric motor + inverter [W]
+  double P_loss_motor =
+      ref_powerLossMap.eval(std::vector<double>{n_motor, M_motor})[0];
+  if(std::isnan(P_loss_motor))
+  {
+    P_loss_motor = 0.0;
+    b_stateValid = false;
+  }
 
-  // Create the model
-  return MMPEVEM(m, r_wheel, Theta, c_rr, c_d, A_front, i_gear, eta_gear, M_max,
-      P_max, M_recup_max, P_recup_max, R_battery, U_battery_0, P_const,
-      powerLossMapString);
+  // Power demand at the battery [W]
+  double P_battery = P_motor + P_loss_motor + P_const;
+  // Total power demand (including the power loss in the battery) [W]
+  double P_total = (U_battery_0 * U_battery_0)/(2.0 * R_battery)
+      - U_battery_0 * std::sqrt((U_battery_0 * U_battery_0
+      - 4.0 * R_battery * P_battery)/(4.0 * R_battery * R_battery));
+
+  ref_powerConsumption = P_total;
+  return b_stateValid;
 }
+
 
 
 
@@ -154,40 +173,10 @@ MMPEVEM loadModelFromFile(const std::string& ref_modelFile)
  * \brief Constructor
  */
 HelpersMMPEVEM::HelpersMMPEVEM()
- : PollutantsInterface::Helper("MMPEVEM", MMPEVEM_BASE, -1),
-   myIndex(MMPEVEM_BASE)
+ : PollutantsInterface::Helper("MMPEVEM", MMPEVEM_BASE, -1)
 { }
 
 
-
-/**
- * \brief Get the emission class that is associated with a name.
- *
- * \param[in] ref_eClass Name of an emission class
- * \param[in] vc The vehicle class to use when determining the default class
- */
-SUMOEmissionClass HelpersMMPEVEM::getClassByName(const std::string& ref_eClass,
-    const SUMOVehicleClass vc)
-{
-  std::string eClass = StringUtils::to_lower_case(ref_eClass);
-
-  if(eClass == "unknown" && !myEmissionClassStrings.hasString("unknown"))
-    myEmissionClassStrings.addAlias("unknown", getClassByName("vw_id3", vc));
-  if(eClass == "default" && !myEmissionClassStrings.hasString("default"))
-    myEmissionClassStrings.addAlias("default", getClassByName("vw_id3", vc));
-  if(myEmissionClassStrings.hasString(eClass))
-    return myEmissionClassStrings.get(eClass);
-
-  int currentIndex = myIndex;
-  myIndex++;
-
-  std::string modelFile = std::string(getenv("SUMO_HOME"))
-      + "/data/emissions/MMPEVEM/" + eClass + ".dat";
-  myModels.insert({currentIndex, loadModelFromFile(modelFile)});
-
-  myEmissionClassStrings.insert(eClass, currentIndex);
-  return currentIndex;
-}
 
 
 
@@ -202,20 +191,65 @@ SUMOEmissionClass HelpersMMPEVEM::getClassByName(const std::string& ref_eClass,
  * \param[in] v Current vehicle velocity [m/s]
  * \param[in] a Current acceleration of the vehicle [m/s^2]
  * \param[in] slope Slope of the road at the vehicle's current position [deg]
+ * \param[in] ptr_energyParams Vehicle parameters
  *
  * \returns The electric power consumption [Wh/s] or 0 for all other emission
  *          types
  */
 double HelpersMMPEVEM::compute(const SUMOEmissionClass c,
     const PollutantsInterface::EmissionType e, const double v, const double a,
-    const double slope, const EnergyParams* /* ptr_energyParams */) const
+    const double slope, const EnergyParams* ptr_energyParams) const
 {
   if(e != PollutantsInterface::ELEC)
     return 0.0;
 
+  // Extract all required parameters
+  // Vehicle mass
+  const double m = ptr_energyParams->getDouble(SUMO_ATTR_VEHICLEMASS);
+  // Wheel radius
+  const double r_wheel = ptr_energyParams->getDouble(SUMO_ATTR_WHEELRADIUS);
+  // Internal moment of inertia
+  const double Theta
+      = ptr_energyParams->getDouble(SUMO_ATTR_INTERNALMOMENTOFINERTIA);
+  // Rolling resistance coefficient
+  const double c_rr
+      = ptr_energyParams->getDouble(SUMO_ATTR_ROLLDRAGCOEFFICIENT);
+  // Air drag coefficient
+  const double c_d = ptr_energyParams->getDouble(SUMO_ATTR_AIRDRAGCOEFFICIENT);
+  // Cross-sectional area of the front of the car
+  const double A_front
+      = ptr_energyParams->getDouble(SUMO_ATTR_FRONTSURFACEAREA);
+  // Gear ratio
+  const double i_gear = ptr_energyParams->getDouble(SUMO_ATTR_GEARRATIO);
+  // Gearbox efficiency
+  const double eta_gear = ptr_energyParams->getDouble(SUMO_ATTR_GEAREFFICIENCY);
+  // Maximum torque
+  const double M_max = ptr_energyParams->getDouble(SUMO_ATTR_MAXIMUMTORQUE);
+  // Maximum power
+  const double P_max = ptr_energyParams->getDouble(SUMO_ATTR_MAXIMUMPOWER);
+  // Maximum recuperation torque
+  const double M_recup_max
+      = ptr_energyParams->getDouble(SUMO_ATTR_MAXIMUMRECUPERATIONTORQUE);
+  // Maximum recuperation power
+  const double P_recup_max
+      = ptr_energyParams->getDouble(SUMO_ATTR_MAXIMUMRECUPERATIONPOWER);
+  // Internal battery resistance
+  const double R_battery
+      = ptr_energyParams->getDouble(SUMO_ATTR_INTERNALBATTERYRESISTANCE);
+  // Nominal battery voltage
+  const double U_battery_0
+      = ptr_energyParams->getDouble(SUMO_ATTR_NOMINALBATTERYVOLTAGE);
+  // Constant power consumption
+  const double P_const
+      = ptr_energyParams->getDouble(SUMO_ATTR_CONSTANTPOWERINTAKE);
+  // Power loss map
+  const CharacteristicMap& ref_powerLossMap
+      = ptr_energyParams->getCharacteristicMap(SUMO_ATTR_POWERLOSSMAP);
+
   double P = 0.0;  // [W]
-  const MMPEVEM& ref_model = myModels.find(c)->second;
-  bool b_stateValid = ref_model.calcPowerConsumption(TS, v, a, slope, P);
+  bool b_stateValid = calcPowerConsumption(m, r_wheel, Theta, c_rr, c_d,
+      A_front, i_gear, eta_gear, M_max, P_max, M_recup_max, P_recup_max,
+      R_battery, U_battery_0, P_const, ref_powerLossMap, TS, v, a, slope, P);
   P /= 3600.0;  // [Wh/s]
 
   if(b_stateValid)
